requirements.txt
yt-dlp
pydub
ffmpeg-python
openai
pinecone
flask
tiktoken
nltk
markdown
python-dotenv